Collapsing
Collapsing Title: The title is larger when the layout is expanded. The text size becomes smaller as the layout is collapsed and scrolled off the screen.
app:layout_scrollFlags: The scroll flags of this layout is typically set to “scroll|exitUntilCollapsed”.
app:collapsedTitleGravity: Specifies the gravity of title in the container when collapsed.
app:contentScrim: This requires specifying a drawable or color value of the CollapsingToolbarLayouts content when it has been scrolled sufficiently off screen eg. ?attr/colorPrimary.
app:scrimAnimationDuration: Specifies the duration used for scrim visibility animations. This requires an integer value such as “100”.